Change projectId and dataset to optional in client
In order for `SanityClienLike` to correspond with the
`SanityClientConfi` type these have to be optional. Otherwise it would
produce type errors in other packages.

Going forward we might be able to solve this better by making the type
of the returned interface change depending on whether you supply
dataset and/or projectId.