feat!: migrate to esm only, unify exports, use pkg-utils for build, switch tests to Vitest (#84)
To support signed URLs (the next PR upstack), I decided to add a specific signed version of the url builder available from a different export path `/signed`. This was to avoid bundling the crypto dependencies we rely on for signing, as this functionality will likely only be used by a very small subset of users.

I thought it would be a good time to "modernize" the repo a little to support that work. That included:

- Replacing `jest`​ with `vitest`​ for tests
- Replacing `microbundle` with `@sanity/pkg-utils` for building
- Updating  the TypeScript configuration to extend `@sanity/pkg-utils` strictest
- Exporting all types defined in `src/types.ts`​ from the default export path

I’ve also added in a `postbuild`​/`prepack`​ step which outputs some files for backwards compatibility. This is to support users who are currently importing types from paths like `@sanity/image-url/lib/types/types`​.