fix(types): export public builder interfaces to avoid TS4094 and correctly name signed type exports (#90)
This fixes a TS4094 error which was a bit of a pain to skirt around.

I had added `ImageUrlBuilder` type exports as anonymous class types, but the classes contained a protected method, which TS says is :no_good:not cool:no_good:.

There's not really a nice way to expose the class type without allowing users to import and instantiate the class itself, which I think we want to avoid. So the solution is to create an `ImageUrlBuilder` interface which the class now `implements`. I've also lifted the protected `constructNewOptions` method out of the class as it didn't really need to be there.

**TL;DR** this gives us a nice `ImageUrlBuilder` type export without any errors, whilst preventing class instantiation.

One other change, which is to differentiate the signed specific type exports. I think still counts as a fix as having identically named types from different export paths with different signatures was IMO wrong. Nobody is using the `/signed` export path yet so this should be fine.