test: add consumer typecheck to catch TS4094 declaration emit regressions (#91)
Added a tiny consumer style type check that runs after build. It recreates the specific setup I've found that triggers a TS4094 error (NodeNext + declaration + exporting a builder instance).

It seems quite edgecase-y, for example this error didn't appear in any of the Sanity repos I updated yesterday, but I suppose there's no harm in adding it. I'd be very interested to hear if there's a better way to check for this!